Able to Upload profile photos to DB
Added way to fetch everytime the user exists the page/app, so that the profile page, opened menus aren't open anymore, and also so that this data is persistent and doesn't reset everytime the user opens/logs into the app. Also able upload profile photos to the DB, and the User object inside the database contians a reference to this profile photo so it can easily be fetched for reference.